Fixed background bug

+++ memory.cpp
@@ -70,7 +70,7 @@ short Memory::readLittleEndian(unsigned short address){
 }
 
 void Memory::OAMDMA(unsigned char highByte){
-    for(unsigned short i=0x00; i<=0xFF; i++){
+    for(unsigned short i=0x00; i<=(unsigned short)0xFF; i++){
         printf("%d\n", i);
         char data = readAddress((highByte<<8) | i); //Check that OAM DMA increases OAM Addresses
         writeAddress(0x2004, data);
